[MOB-2831] Split FeedbackView in multiple components
