+++ firefox-ios/Ecosia/UI/FeedbackView/FeedbackType.swift
@@ -1,0 +1,37 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+import Common
+
+/// FeedbackType represents the type of feedback a user can submit
+enum FeedbackType: String, CaseIterable, Identifiable {
+    case reportIssue = "Report an issue"
+    case generalQuestion = "General question"
+    case suggestionOrFeedback = "Suggestion or feedback"
+
+    var id: String { self.rawValue }
+
+    var analyticsIdentfier: String {
+        switch self {
+        case .reportIssue:
+            return "report_issue"
+        case .generalQuestion:
+            return "general_question"
+        case .suggestionOrFeedback:
+            return "suggestion_or_feedback"
+        }
+    }
+
+    var localizedString: String {
+        switch self {
+        case .reportIssue:
+            return String.localized(.reportIssue)
+        case .generalQuestion:
+            return String.localized(.generalQuestion)
+        case .suggestionOrFeedback:
+            return String.localized(.suggestionOrFeedback)
+        }
+    }
+}
