+++ firefox-ios/Ecosia/UI/FeedbackView/FeedbackView.swift
@@ -1,0 +1,113 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import SwiftUI
+import UIKit
+import Common
+
+/// The SwiftUI view for collecting user feedback
+public struct FeedbackView: View {
+    // User input state
+    @State private var selectedFeedbackType: FeedbackType?
+    @State private var feedbackText: String = ""
+    @State private var isButtonEnabled: Bool = false
+
+    // Theme handling
+    @StateObject private var viewModel = FeedbackViewModel()
+    let windowUUID: WindowUUID?
+
+    // Define a dismiss callback that will be injected by the hosting controller
+    var onDismiss: (() -> Void)?
+    // Callback for notifying when feedback was submitted
+    var onFeedbackSubmitted: (() -> Void)?
+
+    // Layout constants
+    struct UX {
+        static let cornerRadius: CGFloat = .ecosia.borderRadius._l
+        static let buttonCornerRadius: CGFloat = 25
+        static let textEditorHeight: CGFloat = 200
+    }
+
+    public init(windowUUID: WindowUUID? = nil,
+                initialTheme: Theme? = nil) {
+        self.windowUUID = windowUUID
+
+        // Apply initial theme if provided
+        if let theme = initialTheme {
+            _viewModel = StateObject(wrappedValue: FeedbackViewModel(theme: theme))
+        }
+    }
+
+    public var body: some View {
+        NavigationView {
+            ZStack {
+
+                viewModel.backgroundColor.ignoresSafeArea()
+
+                FeedbackContentView(
+                    viewModel: viewModel,
+                    selectedFeedbackType: $selectedFeedbackType,
+                    feedbackText: $feedbackText,
+                    isButtonEnabled: $isButtonEnabled,
+                    updateButtonState: updateButtonState,
+                    sendFeedback: sendFeedback
+                )
+                .navigationTitle(String.localized(.sendFeedback))
+                .navigationBarTitleDisplayMode(.inline)
+                .toolbar {
+                    ToolbarItem(placement: .navigationBarTrailing) {
+                        Button(String.localized(.close)) {
+                            dismiss()
+                        }
+                        .foregroundColor(viewModel.brandPrimaryColor)
+                        .accessibilityIdentifier("close_feedback_button")
+                    }
+                }
+            }
+            .onReceive(NotificationCenter.default.publisher(for: .ThemeDidChange)) { notification in
+                guard let uuid = notification.windowUUID, uuid == windowUUID else { return }
+                let themeManager = AppContainer.shared.resolve() as ThemeManager
+                viewModel.applyTheme(theme: themeManager.getCurrentTheme(for: windowUUID))
+            }
+        }
+    }
+
+    /// Update the state of the send button based on user input
+    private func updateButtonState() {
+        isButtonEnabled = selectedFeedbackType != nil && !feedbackText.isEmpty
+    }
+
+    /// Dismiss the view
+    private func dismiss() {
+        onDismiss?()
+    }
+
+    /// Send the feedback to analytics and dismiss the view
+    private func sendFeedback() {
+        // Gather system information to include in analytics event
+        let deviceType = UIDevice.current.model
+        let operatingSystem = "\(UIDevice.current.systemName) \(UIDevice.current.systemVersion)"
+        let idiom = UIDevice.current.userInterfaceIdiom == .pad ? "iPadOS" : "iOS"
+        let browserVersion = "Ecosia \(idiom) \(Bundle.version)"
+
+        // Send the feedback using the navigation method with the collected data
+        Analytics.shared.sendFeedback([
+            "feedback_type": selectedFeedbackType?.analyticsIdentfier ?? "",
+            "device_type": deviceType,
+            "os": operatingSystem,
+            "browser_version": browserVersion,
+            "feedback_text": feedbackText
+        ])
+
+        // Notify that feedback was submitted
+        onFeedbackSubmitted?()
+
+        // Dismiss the view
+        dismiss()
+    }
+}
+
+#Preview {
+    FeedbackView(windowUUID: UUID())
+}
